prai-roadshow-lab-1-starter/agents/judge/requirements.txt
google-adk[fastapi]
pydantic
uvicorn
fastapi
